Reorganize config of zsh.
I think there's still some refactoring to do, but all PATH related
things will now live in zprofile and aliases/interactive settings will
be set in zshrc. I think there are still some redudancies due to me
collapsing a few files into what is now my zshrc, but for now it seems
to be working. Should go through it again, but need to commit something
in the meantime because it has been detrimental working without it.